RDCC-992: Structure Test Packages To Match Main Packages
