+++ src/main/java/uk/gov/hmcts/reform/profilesync/constants/ProfileSyncEnum.java
@@ -1,0 +1,6 @@
+package uk.gov.hmcts.reform.profilesync.constants;
+
+public enum ProfileSyncEnum {
+    BASIC,
+   BEARER,
+}
