+++ src/main/java/uk/gov/hmcts/reform/profilesync/constants/Source.java
@@ -1,0 +1,5 @@
+package uk.gov.hmcts.reform.profilesync.constants;
+
+public enum Source {
+    API,SYNC
+}
