--- src/main/java/uk/gov/hmcts/reform/profilesync/domain/IdamStatus.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.profilesync.domain;
-
-public enum IdamStatus {
-    ACTIVE,PENDING,SUSPENDED
-}
