--- src/main/java/uk/gov/hmcts/reform/profilesync/domain/Source.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.profilesync.domain;
-
-public enum Source {
-    API,SYNC
-}
