--- src/main/java/uk/gov/hmcts/reform/profilesync/service/ProfileSyncEnum.java
@@ -1,6 +1,0 @@
-package uk.gov.hmcts.reform.profilesync.service;
-
-public enum ProfileSyncEnum {
-    BASIC,
-   BEARER,
-}
