+++ src/main/java/uk/gov/hmcts/reform/profilesync/service/impl/UserAcquisitionServiceImpl.java
@@ -12,11 +12,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.profilesync.client.UserProfileClient;
-import uk.gov.hmcts.reform.profilesync.domain.ErrorResponse;
-import uk.gov.hmcts.reform.profilesync.domain.GetUserProfileResponse;
 import uk.gov.hmcts.reform.profilesync.domain.UserProfileSyncException;
+import uk.gov.hmcts.reform.profilesync.domain.response.ErrorResponse;
+import uk.gov.hmcts.reform.profilesync.domain.response.GetUserProfileResponse;
 import uk.gov.hmcts.reform.profilesync.service.UserAcquisitionService;
-import uk.gov.hmcts.reform.profilesync.util.JsonFeignResponseHelper;
+import uk.gov.hmcts.reform.profilesync.util.JsonFeignResponseUtil;
 
 @Slf4j
 @AllArgsConstructor
@@ -32,7 +32,7 @@ public class UserAcquisitionServiceImpl implements UserAcquisitionService {
         try (Response response = userProfileClient.findUser(bearerToken, s2sToken, id)) {
 
             Class clazz = response.status() > 200 ? ErrorResponse.class : GetUserProfileResponse.class;
-            ResponseEntity responseEntity = JsonFeignResponseHelper.toResponseEntity(response, clazz);
+            ResponseEntity responseEntity = JsonFeignResponseUtil.toResponseEntity(response, clazz);
 
             if (response.status() == 400) {
