+++ src/test/java/uk/gov/hmcts/reform/profilesync/service/impl/UserAcquisitionServiceImplTest.java
@@ -17,10 +17,10 @@ import java.util.UUID;
 import org.junit.Test;
 import org.mockito.Mockito;
 import uk.gov.hmcts.reform.profilesync.client.UserProfileClient;
-import uk.gov.hmcts.reform.profilesync.domain.GetUserProfileResponse;
-import uk.gov.hmcts.reform.profilesync.domain.IdamStatus;
+import uk.gov.hmcts.reform.profilesync.constants.IdamStatus;
 import uk.gov.hmcts.reform.profilesync.domain.UserProfile;
 import uk.gov.hmcts.reform.profilesync.domain.UserProfileSyncException;
+import uk.gov.hmcts.reform.profilesync.domain.response.GetUserProfileResponse;
 import uk.gov.hmcts.reform.profilesync.helper.MockDataProvider;
 import uk.gov.hmcts.reform.profilesync.service.UserAcquisitionService;
